# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/books":{"get":{"operationId":"BookController_getAllBooks","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"BookController_createBook","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookDto"}}}},"responses":{"201":{"description":""}}}},"/books/{id}":{"get":{"operationId":"BookController_getBook","parameters":[],"responses":{"200":{"description":"Returns the book with the given id.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"404":{"description":"Book not found."}}},"put":{"operationId":"BookController_updateBook","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBookDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"BookController_deleteBook","parameters":[],"responses":{"200":{"description":""}}}},"/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDto"}}}},"responses":{"201":{"description":""}}}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}}}}},"info":{"title":"Book Store example","description":"The book store API description","version":"1.0","contact":{}},"tags":[{"name":"bookstore","description":""}],"servers":[],"components":{"schemas":{"Book":{"type":"object","properties":{}},"User":{"type":"object","properties":{}},"CreateBookDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"author":{"type":"string"},"price":{"type":"number"},"category":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["title","description","author","price","category","user"]},"UpdateBookDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"author":{"type":"string"},"price":{"type":"number"},"category":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["title","description","author","price","category","user"]},"SignupDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{}}}}}